React refactor: New dir structure|some new content
Began introducing a more React-typical directory structure.
Added `/src/lib/` for holding various bits of functionality that the eventual React components may use.
Moved most existing functionality into `/src/lib/`. Will sort these out later.

Deleted long-unused files from `src/currently_unused/`.

Moved schemas to `src/model_json/schemas/`. `.gitmodules` will reflect this change.

Added some new content from https://github.com/quantellia/di-controls/tree/controls-schema-integration
- merged .gitignore contents
- Moved example model JSON into `src/model_json`. These new model JSON files are more fully-featured, with controls and scripts attached.

Replaced references to `schema_compliant_cdd.json` with `src/model_json/coffee.json`.
Deleted `schema_compliant_cdd.json`

Updated references for moved files so that the Authoring Tool is still currently functional with the old JointJS setup.
