+++ src/model_json/basic_adder.json
@@ -1,0 +1,242 @@
+{
+  "$schema": "./schemas/Causal-Decision-Model.json",
+  "meta": {
+    "uuid": "28eefdc9-d38c-458d-a0b1-cba3706a09e4",
+    "name": "Example/Test Model",
+    "summary": "An example CDM for testing/demonstrating the CDM JSON schema.",
+    "status": "Draft",
+    "version": "0.0.1",
+    "creator": "9744a75e-6339-43ad-beec-b1a4f23e7e26",
+    "createdDate": "2024-08-20T09:45:41-04:00"
+  },
+  "runnableModels": [
+    {
+      "meta": {
+        "uuid": "c7baceb4-33cf-4b7a-acea-4cd5e15340e5",
+        "name": "Example evaluatable causal decision model."
+      },
+      "elements": [
+        {
+          "meta": {
+            "uuid": "8285d781-826f-4de3-9ded-be75e46935d3",
+            "name": "Sum inputs"
+          },
+          "inputs": [
+            "0dd65d95-d90a-454e-9489-34237d740afd",
+            "b5a14d60-3abc-4306-bc12-3c0678851279"
+          ],
+          "outputs": [
+            "74c6b8cf-ea6b-4ae0-bc52-dc67e0efe630"
+          ],
+          "functionName": "add",
+          "evaluatableAsset": "a14ae3d2-c070-4485-b6d1-3af8e1a87cad"
+        }
+      ]
+    }
+  ],
+  "evaluatableAssets": [
+      {
+          "meta": {
+              "uuid": "a14ae3d2-c070-4485-b6d1-3af8e1a87cad",
+              "name": "Add script"
+          },
+          "evalType": "Script",
+          "content": {
+              "script": "KGZ1bmN0aW9uICgpIHsKICBjb25zdCBhZGQgPSBmdW5jdGlvbiAodmFscykgewogICAgY29uc29sZS5sb2coIkFERCBSVU5OSU5HISEhISIpOwogICAgbGV0IHN1bSA9IDA7CiAgICB2YWxzLmZvckVhY2goKHZhbCkgPT4gewogICAgICBzdW0gKz0gdmFsOwogICAgfSk7CiAgICByZXR1cm4gW3N1bV07CiAgfTsKCiAgcmV0dXJuIHsgZnVuY01hcDogeyAiYWRkIjogYWRkIH0gfTsKfSkoKTs=",
+              "language": "javascript"
+          }
+      }
+  ],
+  "inputOutputValues": [
+      {
+          "meta": {
+              "uuid": "0dd65d95-d90a-454e-9489-34237d740afd",
+              "title": "Add_Input_1"
+          },
+          "data": 30
+      },
+      {
+          "meta": {
+              "uuid": "b5a14d60-3abc-4306-bc12-3c0678851279",
+              "title": "Add_Input_2"
+          },
+          "data": 27
+      },
+      {
+          "meta": {
+              "uuid": "74c6b8cf-ea6b-4ae0-bc52-dc67e0efe630",
+              "title": "Add_Output"
+          },
+          "data": null
+      }
+  ],
+  "diagrams": [
+    {
+      "meta": {
+        "uuid": "999cad87-3323-4919-b870-1b91170abe0d",
+        "name": "Example diagram associated with the causal decision model."
+      },
+      "elements": [
+        {
+          "meta": {
+            "uuid": "b7dac5d9-5c33-45bd-b30e-752f04997c1f",
+            "name": "A constraint"
+          },
+          "causalType": "CUSTOM_Constraint",
+          "position": {
+            "x": 50,
+            "y": 510
+          }
+        },
+        {
+          "meta": {
+            "uuid": "15e06e39-f74d-4e83-aed2-817106102c1a",
+            "name": "Range For Add Input 1"
+          },
+          "causalType": "Lever",
+          "position": {
+            "x": 120,
+            "y": 85
+          },
+          "displays": [
+            {
+              "meta": {
+                "uuid": "8392886a-84ab-4623-9b53-2c571b88ddca"
+              },
+              "displayType": "controlRange",
+              "content": {
+                "controlParameters": {
+                  "min": 0,
+                  "max": 100,
+                  "step": 1,
+                  "value": -1,
+                  "isInteractive": true
+                }
+              }
+            }
+          ]
+        },
+        {
+          "meta": {
+            "uuid": "200ccb9a-bc2e-41ce-9e36-5b5cee247b5d",
+            "name": "Range For Add Input 2"
+          },
+          "causalType": "Lever",
+          "position": {
+            "x": 120,
+            "y": 250
+          },
+          "displays": [
+            {
+              "meta": {
+                "uuid": "83639f65-0fb8-4ec2-a031-7a3c5238d6f6"
+              },
+              "displayType": "controlRange",
+              "content": {
+                "controlParameters": {
+                  "min": 0,
+                  "max": 100,
+                  "step": 1,
+                  "value": -1,
+                  "isInteractive": true
+                }
+              }
+            }
+          ]
+        },
+        {
+          "meta": {
+            "uuid": "eaf1c6d3-0797-46fa-b005-52f7ee23c8f0",
+            "name": "Range For Sum"
+          },
+          "causalType": "Outcome",
+          "position": {
+            "x": 560,
+            "y": 150
+          },
+          "displays": [
+            {
+              "meta": {
+                "uuid": "0214a5a2-8379-4aa0-94cb-776bdcc40738"
+              },
+              "displayType": "controlRange",
+              "content": {
+                "controlParameters": {
+                  "min": 0,
+                  "max": 200,
+                  "step": 1,
+                  "value": -1,
+                  "isInteractive": false
+                }
+              }
+            }
+          ]
+        },
+        {
+          "meta": {
+            "uuid": "4feca1ad-8ebd-4817-852b-03ae277213c9",
+            "name": "Diagram-only Element",
+            "description": "This element is not related to any evaluatable assets via any controls. It also has a null causal type."
+          },
+          "causalType": "Intermediate",
+          "position": {
+            "x": 650,
+            "y": 0
+          }
+        }
+      ],
+      "dependencies": [
+        {
+          "meta": {
+            "uuid": "907d9ac7-547d-456a-8622-f311da1e8fa8",
+            "name": "Add Input 1 --> Sum"
+          },
+          "source": "15e06e39-f74d-4e83-aed2-817106102c1a",
+          "target": "eaf1c6d3-0797-46fa-b005-52f7ee23c8f0"
+        },
+        {
+          "meta": {
+            "uuid": "a1e4006a-e814-435a-a9c7-ff8305efdadf",
+            "name": "Add Input 2 --> Sum"
+          },
+          "source": "200ccb9a-bc2e-41ce-9e36-5b5cee247b5d",
+          "target": "eaf1c6d3-0797-46fa-b005-52f7ee23c8f0"
+        },
+        {
+          "meta": {
+            "uuid": "baa19a4d-c194-404c-8e51-9671a1c3fe3a",
+            "name": "Add Input 1 --> Diagram-only Element"
+          },
+          "source": "15e06e39-f74d-4e83-aed2-817106102c1a",
+          "target": "4feca1ad-8ebd-4817-852b-03ae277213c9"
+        }
+      ]
+    }
+  ],
+  "controls": [
+    {
+      "meta": {
+        "uuid": "94e5a529-ca27-4c1f-92a7-0c5fafa4fe18",
+        "name": "Control: Add Input 1"
+      },
+      "inputOutputValues": ["0dd65d95-d90a-454e-9489-34237d740afd"],
+      "displays": ["8392886a-84ab-4623-9b53-2c571b88ddca"]
+    },
+    {
+      "meta": {
+        "uuid": "96c8f736-dc2f-4302-9da3-b249b8045e7b",
+        "name": "Control: Add Input 2"
+      },
+      "inputOutputValues": ["b5a14d60-3abc-4306-bc12-3c0678851279"],
+      "displays": ["83639f65-0fb8-4ec2-a031-7a3c5238d6f6"]
+    },
+    {
+      "meta": {
+        "uuid": "c08f2810-9825-4fc8-946c-85d77d61a960",
+        "name": "Control: Add Output"
+      },
+      "inputOutputValues": ["74c6b8cf-ea6b-4ae0-bc52-dc67e0efe630"],
+      "displays": ["0214a5a2-8379-4aa0-94cb-776bdcc40738"]
+    }
+  ]
+}
