+++ src/uiButtons.js
@@ -1,6 +1,6 @@
 import * as joint from "@joint/core/dist/joint.js";
 import {v4 as uuidv4} from 'uuid';
-import * as fileIO from "./fileIO.js";
+import * as fileIO from "./lib/fileIO.js";
 
 /**
  * Defines function button object. Used for clickable buttons that
